Script finally working after 6 hours

+++ Part3/Part1/LPTokens.sol
@@ -7,14 +7,14 @@ contract LPToken {
     string constant name = "LPToken";
     string constant symbol = "LPT";
     address internal DexAddr;
-    uint internal totalTokens = 0;
-    mapping (address => uint) internal _balance;
+    uint256 internal totalTokens = 0;
+    mapping (address => uint256) internal _balance;
 
     constructor(address DEXAddr){
         DexAddr = DEXAddr;
     }
 
-    function _mint(uint _amount, address receiver) internal returns (uint256) {
+    function _mint(uint256 _amount, address receiver) internal returns (uint256) {
         require(_amount != 0);
         _balance[receiver]+= _amount;
         totalTokens += _amount;
@@ -25,12 +25,12 @@ contract LPToken {
        return _balance[account];
     }
 
-    function generateTokens(uint _amount, address receiver) public returns (bool success, uint) {
+    function generateTokens(uint256 _amount, address receiver) public returns (bool success, uint256) {
         if (msg.sender != DexAddr) return (false, _balance[receiver]);
         return (true, _mint(_amount, receiver));
     }
 
-    function burn(uint _amount, address owner) public returns (bool success, uint) {
+    function burn(uint256 _amount, address owner) public returns (bool success, uint256) {
         if(msg.sender != DexAddr) return (false, _balance[owner]);
         if (_balance[owner] < _amount) return (false, _balance[owner]);
         _balance[owner] -= _amount;
