+++ Part3/Part1/token.sol
@@ -4,7 +4,7 @@ pragma solidity ^0.8.20;
 import "@openzeppelin/contracts/token/ERC20/ERC20.sol";
 
 contract MyToken is ERC20 {
-    uint constant _total_supply = 100*(10**18);
+    uint256 constant _total_supply = 100*(10**18);
     constructor(string memory name, string memory symbol) ERC20(name, symbol) {
         _mint(msg.sender, _total_supply);
     }
